Rename ObjectDetailsDrawer to ObjectDetailsPanel
